+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.3] - 2026-01-01
+
 ### Fixed
 - Fix `gcx setup` failing on first run when config directory doesn't exist
+- Fix yq syntax error when adding organizations/identities with dynamic names
+- Fix gum choose failing when no kubeconfig or ADC credential files exist
 
 ## [1.2.2] - 2026-01-01
 
@@ -49,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multi-organization, multi-identity support
 - Config export/import for team sharing
 
-[Unreleased]: https://github.com/cirrus-tools/gcx/compare/v1.2.2...HEAD
+[Unreleased]: https://github.com/cirrus-tools/gcx/compare/v1.2.3...HEAD
+[1.2.3]: https://github.com/cirrus-tools/gcx/compare/v1.2.2...v1.2.3
 [1.2.2]: https://github.com/cirrus-tools/gcx/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/cirrus-tools/gcx/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/cirrus-tools/gcx/compare/v1.1.0...v1.2.0